+++ src/nb/spi.rs
@@ -13,10 +13,10 @@
 ///
 /// - Data is only guaranteed to be clocked out when the `read` call succeeds.
 /// The slave select line shouldn't be released before that.
-///
-/// - Some SPIs can work with 8-bit *and* 16-bit words. You can overload this trait with different
-/// `Word` types to allow operation in both modes.
-pub trait FullDuplex<Word> {
+pub trait FullDuplex {
+    /// Word type
+    type Word;
+
     /// An enumeration of SPI errors
     type Error;
 
@@ -24,10 +24,10 @@ pub trait FullDuplex<Word> {
     ///
     /// **NOTE** A word must be sent to the slave before attempting to call this
     /// method.
-    fn read(&mut self) -> nb::Result<Word, Self::Error>;
+    fn read(&mut self) -> nb::Result<Self::Word, Self::Error>;
 
     /// Writes a word to the slave
-    fn write(&mut self, word: Word) -> nb::Result<(), Self::Error>;
+    fn write(&mut self, word: Self::Word) -> nb::Result<(), Self::Error>;
 }
 
 /// Clock polarity